First work on S3 Express Directory Bucket
Separate S3 express example as there are some s3 express specific
parameters required.

Addtionally this requires a specialized Docker image to upgrade Hadoop
to 3.4.0 in order to suppport s3-express.

Hadoop 3.4.0 is a big deal to upgrade as it brings in many new
dependencies such as the AWS SDK Java V2 library.